aws/ecs_service_connect/docker/rekog/requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
boto3